+++ bscampp/functions.py
@@ -40,18 +40,21 @@ def readData(workdir, dry_run=False):
     if Configs.qaln_path is not None: 
         ref_dict = utils.read_data(Configs.aln_path)
         q_dict = utils.read_data(Configs.qaln_path)
-        aln_path, qaln_path = Configs.aln_path, Configs.qaln_path
+        #aln_path, qaln_path = Configs.aln_path, Configs.qaln_path
     else:
         aln_dict = utils.read_data(Configs.aln_path)
         ref_dict, q_dict = utils.seperate(aln_dict, leaf_dict)
 
-        # after separating queries from the reference alignment, write
-        # them to to TEMP/
-        qaln_path = os.path.join(workdir, 'qaln.fa')
-        write_fasta(qaln_path, q_dict)
-        
-        aln_path = os.path.join(workdir, 'aln.fa')
-        write_fasta(aln_path, ref_dict)
+    # after separating queries from the reference alignment, write
+    # them to to TEMP/
+    # Updated on 3.5.2025 by Chengze Shen
+    #   - regardless of the input choices, write a copy of both reference
+    #     and query alignment to the workdir
+    qaln_path = os.path.join(workdir, 'qaln.fa')
+    write_fasta(qaln_path, q_dict)
+    
+    aln_path = os.path.join(workdir, 'aln.fa')
+    write_fasta(aln_path, ref_dict)
 
     t1 = time.perf_counter()
     _LOG.info('Time to read in input data: {} seconds'.format(t1 - t0))